Make bravewength package organization more consistent
